requirements.txt
celery==5.5.3
flower==2.0.1
redis==7.0.1
celery==5.5.3
django-celery-beat==2.8.1
django_celery_results==2.6.0
flower==2.0.1
redis==7.0.1
